Size hashtable according to device RAM.
Right now we always use the default hashtable size of Stockfish, which
is 16MB. A modern, multiprocessor phone is very fast and will fill such
a hashtable in seconds, making further search much less efficient.

Determine the device RAM when initializing the plugin and allocate the
hashtable to be 1/16th the usable device RAM.

For a 4GB RAM Android device, this means using about ~220MB hashtables.
For a 1GB one, this means about ~55MB, which seems reasonable.